Take out internal class and interfaces.
